data-products/customer360/tools/requirements.txt
PyYAML>=6.0.1
jinja2>=3.1.3
python-dotenv>=1.0.1
